Weather: remove WeatherStation
Trafikverket has indicated the WeatherStation will be removed from their
API on the 30th of January 2024. This removes the type and updates the
example to use WeatherMeasurepoint instead.

Some of the .Value fields in the XML are tagged as decimal, but were
incorrectly translated to int in Go. Though float64 isn't quite the
equivalent for decimal, it's good enough for our uses here.