Make Request.Build() do some sanity checking
When calling .Build() an error will now be returned if no API key was
set. Without the API key, the request would fail.

In that same spirit, a Request with no Queries makes no sense, so that's
blocked too.

It's still possbile to manually xml.Marshal a Request instead of calling
Build(), at which point no validation would occur.